# Fix NestedLoopJoin performance regression